requirements.txt
requests
pyjwt
pycrypto
